+++ src/api/core/health.rs
@@ -11,10 +11,10 @@ use serde_derive::{Deserialize, Serialize};
 use serde_with::{serde_as, skip_serializing_none};
 #[cfg(feature = "sidekiq")]
 use sidekiq::redis_rs::cmd;
-#[cfg(feature = "sidekiq")]
+#[cfg(any(feature = "db-sql", feature = "sidekiq"))]
 use std::time::Duration;
 use std::time::Instant;
-#[cfg(feature = "sidekiq")]
+#[cfg(any(feature = "db-sql", feature = "sidekiq"))]
 use tokio::time::timeout;
 use tracing::instrument;
 
@@ -79,35 +79,21 @@ where
     S: Clone + Send + Sync + 'static,
 {
     let timer = Instant::now();
-    #[cfg(any(feature = "sidekiq", feature = "db-sql"))]
-    #[cfg(feature = "db-sql")]
-    let db = {
-        let db_timer = Instant::now();
-        let db_status = if ping_db(state.db()).await.is_ok() {
-            Status::Ok
-        } else {
-            Status::Err(ErrorData { msg: None })
-        };
-        let db_timer = db_timer.elapsed();
-        ResourceHealth {
-            status: db_status,
-            acquire_conn_latency: None,
-            ping_latency: None,
-            latency: db_timer.as_millis(),
-        }
-    };
 
-    #[cfg(feature = "sidekiq")]
-    let (redis_enqueue, redis_fetch) = {
-        let redis_enqueue = redis_health(state.redis_enqueue());
-        if let Some(redis_fetch) = state.redis_fetch() {
-            let (redis_enqueue, redis_fetch) =
-                tokio::join!(redis_enqueue, redis_health(redis_fetch));
-            (redis_enqueue, Some(redis_fetch))
-        } else {
-            (redis_enqueue.await, None)
-        }
-    };
+    #[cfg(any(feature = "db-sql", feature = "sidekiq"))]
+    let timeout_duration = Duration::from_secs(1);
+
+    #[cfg(all(feature = "db-sql", feature = "sidekiq"))]
+    let (db, (redis_enqueue, redis_fetch)) = tokio::join!(
+        db_health(state, timeout_duration),
+        all_redis_health(state, timeout_duration)
+    );
+
+    #[cfg(all(feature = "db-sql", not(feature = "sidekiq")))]
+    let db = db_health(state, timeout_duration).await;
+
+    #[cfg(all(not(feature = "db-sql"), feature = "sidekiq"))]
+    let (redis_enqueue, redis_fetch) = all_redis_health(state, timeout_duration).await;
 
     Ok(HeathCheckResponse {
         latency: timer.elapsed().as_millis(),
@@ -121,17 +107,59 @@ where
 }
 
 #[cfg(feature = "db-sql")]
+pub(crate) async fn db_health<S>(state: &AppContext<S>, duration: Duration) -> ResourceHealth
+where
+    S: Clone + Send + Sync + 'static,
+{
+    let db_timer = Instant::now();
+    let db_status = match ping_db(state.db(), duration).await {
+        Ok(_) => Status::Ok,
+        Err(err) => Status::Err(ErrorData {
+            msg: Some(err.to_string()),
+        }),
+    };
+    let db_timer = db_timer.elapsed();
+    ResourceHealth {
+        status: db_status,
+        acquire_conn_latency: None,
+        ping_latency: None,
+        latency: db_timer.as_millis(),
+    }
+}
+
+#[cfg(feature = "db-sql")]
 #[instrument(skip_all)]
-async fn ping_db(db: &DatabaseConnection) -> RoadsterResult<()> {
-    db.ping().await?;
+async fn ping_db(db: &DatabaseConnection, duration: Duration) -> RoadsterResult<()> {
+    timeout(duration, db.ping()).await??;
     Ok(())
 }
 
 #[cfg(feature = "sidekiq")]
+pub(crate) async fn all_redis_health<S>(
+    state: &AppContext<S>,
+    duration: Duration,
+) -> (ResourceHealth, Option<ResourceHealth>)
+where
+    S: Clone + Send + Sync + 'static,
+{
+    {
+        let redis_enqueue = redis_health(state.redis_enqueue(), duration);
+        if let Some(redis_fetch) = state.redis_fetch() {
+            let (redis_enqueue, redis_fetch) =
+                tokio::join!(redis_enqueue, redis_health(redis_fetch, duration));
+            (redis_enqueue, Some(redis_fetch))
+        } else {
+            (redis_enqueue.await, None)
+        }
+    }
+}
+
+#[cfg(feature = "sidekiq")]
 #[instrument(skip_all)]
-async fn redis_health(redis: &sidekiq::RedisPool) -> ResourceHealth {
+async fn redis_health(redis: &sidekiq::RedisPool, duration: Duration) -> ResourceHealth {
     let redis_timer = Instant::now();
-    let (redis_status, acquire_conn_latency, ping_latency) = match ping_redis(redis).await {
+    let (redis_status, acquire_conn_latency, ping_latency) = match ping_redis(redis, duration).await
+    {
         Ok((a, b)) => (Status::Ok, Some(a.as_millis()), Some(b.as_millis())),
         Err(err) => (
             Status::Err(ErrorData {
@@ -152,9 +180,12 @@ async fn redis_health(redis: &sidekiq::RedisPool) -> ResourceHealth {
 
 #[cfg(feature = "sidekiq")]
 #[instrument(skip_all)]
-async fn ping_redis(redis: &sidekiq::RedisPool) -> RoadsterResult<(Duration, Duration)> {
+async fn ping_redis(
+    redis: &sidekiq::RedisPool,
+    duration: Duration,
+) -> RoadsterResult<(Duration, Duration)> {
     let timer = Instant::now();
-    let mut conn = timeout(Duration::from_secs(1), redis.get()).await??;
+    let mut conn = timeout(duration, redis.get()).await??;
     let acquire_conn_latency = timer.elapsed();
 
     let timer = Instant::now();
